requirements.txt
alembic==1.4.2
bcrypt==3.1.7
blinker==1.4
certifi==2020.6.20
cffi==1.14.0
chardet==3.0.4
click==7.1.2
cryptography==2.9.2
Flask==1.1.2
Flask-Bcrypt==0.7.1
Flask-Cors==1.10.3
Flask-Login==0.5.0
Flask-Mail==0.9.1
Flask-Migrate==2.5.3
Flask-SocketIO==4.3.1
Flask-SQLAlchemy==2.4.3
Flask-WTF==0.14.3
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.2
Mako==1.1.3
MarkupSafe==1.1.1
nexmo==2.4.0
numpy==1.19.0
opencv-python==4.2.0.34
pycparser==2.20
PyJWT==1.7.1
pypng==0.0.20
PyQRCode==1.2.1
python-dateutil==2.8.1
python-editor==1.0.4
python-engineio==3.13.1
python-socketio==4.6.0
pytz==2020.1
pyzbar==0.1.8
requests==2.24.0
six==1.15.0
SQLAlchemy==1.3.18
urllib3==1.25.9
Werkzeug==1.0.1
WTForms==2.3.1
alembic==1.4.2
bcrypt==3.1.7
blinker==1.4
certifi==2020.6.20
cffi==1.14.0
chardet==3.0.4
click==7.1.2
cryptography==2.9.2
Flask==1.1.2
Flask-Bcrypt==0.7.1
Flask-Cors==1.10.3
Flask-Login==0.5.0
Flask-Mail==0.9.1
Flask-Migrate==2.5.3
Flask-SocketIO==4.3.1
Flask-SQLAlchemy==2.4.3
Flask-WTF==0.14.3
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.2
Mako==1.1.3
MarkupSafe==1.1.1
nexmo==2.4.0
numpy==1.19.0
opencv-python==4.2.0.34
pycparser==2.20
PyJWT==1.7.1
pypng==0.0.20
PyQRCode==1.2.1
python-dateutil==2.8.1
python-editor==1.0.4
python-engineio==3.13.1
python-socketio==4.6.0
pytz==2020.1
pyzbar==0.1.8
requests==2.24.0
six==1.15.0
SQLAlchemy==1.3.18
urllib3==1.25.9
Werkzeug==1.0.1
WTForms==2.3.1
